% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triact_class.R
\name{Triact}
\alias{Triact}
\title{R6 class for accelerometer data from cows}
\description{
An object for containing and analyzing triaxial accelerometer data from hind legs of cows.
Analyses focus on standing and lying behaviour and activity of the cows.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{Raw and classified accelerometer data. Modified by $load_data, $add_lying, $add_side methods.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-load_data}{\code{Triact$load_data()}}
\item \href{#method-add_lying}{\code{Triact$add_lying()}}
\item \href{#method-add_side}{\code{Triact$add_side()}}
\item \href{#method-extract_liedown}{\code{Triact$extract_liedown()}}
\item \href{#method-extract_standup}{\code{Triact$extract_standup()}}
\item \href{#method-get_activity_by_iterval}{\code{Triact$get_activity_by_iterval()}}
\item \href{#method-get_activity_by_bout}{\code{Triact$get_activity_by_bout()}}
\item \href{#method-clone}{\code{Triact$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_data"></a>}}
\if{latex}{\out{\hypertarget{method-load_data}{}}}
\subsection{Method \code{load_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Triact$load_data(
  input,
  id_substring,
  start_time = NULL,
  end_time = NULL,
  timeXYZ_cols = c(1, 2, 3, 4),
  time_format = NULL,
  tz = Sys.timezone(),
  sep = "auto",
  skip = "__auto__",
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_lying"></a>}}
\if{latex}{\out{\hypertarget{method-add_lying}{}}}
\subsection{Method \code{add_lying()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Triact$add_lying(crit_lie = 0.5, k = 121, check = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_side"></a>}}
\if{latex}{\out{\hypertarget{method-add_side}{}}}
\subsection{Method \code{add_side()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Triact$add_side(crit_left = -0.5)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extract_liedown"></a>}}
\if{latex}{\out{\hypertarget{method-extract_liedown}{}}}
\subsection{Method \code{extract_liedown()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Triact$extract_liedown(sec_before = 0, sec_after = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extract_standup"></a>}}
\if{latex}{\out{\hypertarget{method-extract_standup}{}}}
\subsection{Method \code{extract_standup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Triact$extract_standup(sec_before = 0, sec_after = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_activity_by_iterval"></a>}}
\if{latex}{\out{\hypertarget{method-get_activity_by_iterval}{}}}
\subsection{Method \code{get_activity_by_iterval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Triact$get_activity_by_iterval(interval = "hour", lag_in_s = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_activity_by_bout"></a>}}
\if{latex}{\out{\hypertarget{method-get_activity_by_bout}{}}}
\subsection{Method \code{get_activity_by_bout()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Triact$get_activity_by_bout(bout_type = "all")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Triact$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
